User specific content added

+++ BookAMech.API/BookAMech/Controllers/V1/ReservationsController.cs
@@ -3,14 +3,18 @@ using BookAMech.Contracts.V1.Requests;
 using BookAMech.Contracts.V1.Responses;
 using BookAMech.Domain;
 using BookAMech.Services;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using BookAMech.Extensions;
 
 namespace BookAMech.Controllers.V1
 {
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)] //Everything accessed for this controller needs a valid token
     public class ReservationsController : Controller
     {
         private readonly IReservationService _reservationService;
@@ -29,6 +33,15 @@ namespace BookAMech.Controllers.V1
         [HttpGet(ApiRoutes.Reservations.Get)]
         public async Task<IActionResult> GetReservationAsync([FromRoute] Guid reservationId)
         {
+
+            var userOwnReservation = await _reservationService.userOwnReservationAsync(reservationId, HttpContext.GetUserId());
+
+            //Check for user owns reservation
+            if (!userOwnReservation)
+            {
+                return BadRequest(new { error = "You do not own this reservation" });
+            }
+
             var reservation = await _reservationService.GetReservationByIdAsync(reservationId);
 
             if (reservation == null)
@@ -40,6 +53,15 @@ namespace BookAMech.Controllers.V1
         [HttpPut(ApiRoutes.Reservations.Update)]
         public async Task<IActionResult> UpdateReservationAsync([FromRoute]Guid reservationId, [FromBody] UpdateReservationRequest request)
         {
+
+            var userOwnReservation = await _reservationService.userOwnReservationAsync(reservationId, HttpContext.GetUserId());
+
+            //Check for user owns reservation
+            if (!userOwnReservation)
+            {
+                return BadRequest(new { error = "You do not own this reservation" });
+            }
+
             var reservation = new Reservation
             {
                 Id = reservationId,
@@ -47,8 +69,7 @@ namespace BookAMech.Controllers.V1
                 CompanyName = request.CompanyName,
                 StreetAddress = request.StreetAddress,
                 StreetNumber = request.StreetNumber,
-                Phonenumber = request.Phonenumber
-             
+                Phonenumber = request.Phonenumber       
             };    
 
             var updated = await _reservationService.UpdateReservationAsync(reservation);  
@@ -62,6 +83,14 @@ namespace BookAMech.Controllers.V1
         [HttpDelete(ApiRoutes.Reservations.Delete)]
         public async Task<IActionResult> DeleteReservationAsync([FromRoute] Guid reservationId)
         {
+            var userOwnReservation = await _reservationService.userOwnReservationAsync(reservationId, HttpContext.GetUserId());
+
+            //Check for user owns reservation
+            if (!userOwnReservation)
+            {
+                return BadRequest(new { error = "You do not own this reservation" });
+            }
+
             var deleted = await _reservationService.DeleteReservationAsync(reservationId);
 
             if (deleted)
@@ -75,14 +104,15 @@ namespace BookAMech.Controllers.V1
         public async Task<IActionResult> CreateReservation([FromBody] CreateReservationRequest reservationRequest)
         {
 
-            var reservation = new Reservation 
-            {               
+            var reservation = new Reservation
+            {
                 CustomerName = reservationRequest.CustomerName,
                 CompanyName = reservationRequest.CompanyName,
                 StreetAddress = reservationRequest.StreetAddress,
                 StreetNumber = reservationRequest.StreetNumber,
                 Phonenumber = reservationRequest.Phonenumber,
-                startDate = reservationRequest.startDate
+                startDate = reservationRequest.startDate,
+                UserId = HttpContext.GetUserId()
             };
